+++ device/esp_tinyusb/CHANGELOG.md
@@ -1,6 +1,7 @@
-## [Unreleased]
+## 2.0.1~1
 
-- Fixed forward compatibility with TinyUSB 0.19
+- esp_tinyusb: Claim forward compatibility with TinyUSB 0.19
+- CDC: Added support for new VFS API (for esp-idf v5.4 and higher)
 
 ## 2.0.1